+++ libraries/A02YYUW/A02YYUW.h
@@ -3,34 +3,30 @@
 //    FILE: A02YYUW.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2026-05-07
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 // PURPOSE: Arduino library for A02YYUW serial distance sensor
 //     URL: https://github.com/RobTillaart/A02YYUW
 
 
 #include "Arduino.h"
 
-#define A02YYUW_LIB_VERSION         (F("0.1.1"))
+#define A02YYUW_LIB_VERSION         (F("0.1.2"))
 
-constexpr int A02YY_OK = 0;
-constexpr int A02YY_ERR_CRC = -100;
+constexpr int A02YY_OK          = 0;
+constexpr int A02YY_ERR_CRC     = -100;
+constexpr int A02YY_ERR_TIMEOUT = -101;
 
 
-class A02YYUW
+////////////////////////////////////////////
+//
+//  BASE CLASS
+//
+class A02YY
 {
 public:
-  A02YYUW(Stream * str, uint8_t TX);
+  A02YY(Stream * str);
 
   void     begin();
-
-  //  needed for the UART AUTO only
-  //  true  = processed mode
-  //  false = real time mode (default)
-  void     setProcessingMode(bool mode);
-  bool     getProcessingMode();
-
-  //  needed for the UART CONTROLLED only
-  void     request();
 
   //  needed by all
   bool     newDistance();
@@ -50,6 +46,7 @@ protected:
   uint8_t  _high, _low;
   uint16_t _millimetres;
   uint32_t _lastRead;
+  uint32_t _lastHeader;
 
   int      _error;
 };
@@ -57,30 +54,54 @@ protected:
 
 ////////////////////////////////////////////
 //
-//  DERIVED CLASSES (experimental wrappers)
+//  DERIVED CLASSES - UART AUTO - A02YYU(W)
 //
-class A02YYU : public A02YYUW
+class A02YYU : public A02YY
 {
   public:
-  A02YYU(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYU(Stream * str, uint8_t TX) : A02YY(str)
   {
-  }
+    _TX = TX;
+    //  needed? as Serial implies TX == OUTPUT.
+    pinMode(_TX, OUTPUT);
+  };
+
+  //  true  = processed mode
+  //  false = real time mode (default)
+  void setProcessingMode(bool mode);
+  bool getProcessingMode();
 };
 
-class A02YYT : public A02YYUW
+//  wrapper only
+class A02YYUW : public A02YYU
 {
   public:
-  A02YYT(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYUW(Stream * str, uint8_t TX) : A02YYU(str, TX)
   {
-  }
+  };
 };
 
-class A02YYTW : public A02YYUW
+
+////////////////////////////////////////////
+//
+//  DERIVED CLASSES - UART CONTROLLED - A02YYT(W)
+//
+class A02YYT : public A02YY
 {
   public:
-  A02YYTW(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYT(Stream * str) : A02YY(str)
   {
-  }
+  };
+  void request();
+};
+
+//  wrapper only
+class A02YYTW : public A02YYT
+{
+  public:
+  A02YYTW(Stream * str) : A02YYT(str)
+  {
+  };
 };